allow up/down of LAN interfaces
we're on ESP32 C6 and have LAN via a WizNet W5500 on the SPI bus.
bringing up/down/up the interface during runtime leads to huge
latencies (in a ~1s grid), or leaves LAN defunct.
this adds .deinit() and __del__() to the lan. additionally on
instantiation (get_lan()) we re-instantiate half-initialized
lan objects.

Signed-off-by: matthias wenzel <wenzel@wenzellabs.de>